added new functions for the order page
