+++ shared/components/shared/cart-drawer-item.tsx
@@ -18,12 +18,13 @@ export const CartDrawerItem: React.FC<Props> = ({
   price,
   quantity,
   details,
+  disabled,
   onClickCountButton,
   onClickRemove,
   className,
 }) => {
   return (
-    <div className={cn("flex bg-white p-5 gap-6", className)}>
+    <div className={cn("flex bg-white p-5 gap-6", { "opacity-50 pointer-events-none": disabled }, className)}>
       <CartItem.Image src={imageUrl} />
 
       <div className='flex-1'>
