+++ shared/components/shared/index.ts
@@ -21,5 +21,6 @@ export { CartDrawer } from "./cart-drawer";
 export { CountButton } from "./count-button";
 export { CountIconButton } from "./count-icon-button";
 export { CartDrawerItem } from "./cart-drawer-item";
+export { ProductForm } from "./product-form";
 
 export * from "./modals/index";
