+++ shared/components/shared/modals/choose-product-modal.tsx
@@ -9,6 +9,7 @@ import { ProductWithRelations } from "@/@types/prisma";
 import { ChoosePizzaForm } from "../choose-pizza-form";
 import { useCartStore } from "@/shared/store/cart";
 import toast from "react-hot-toast";
+import { ProductForm } from "../product-form";
 
 interface Props {
   product: ProductWithRelations;
@@ -17,69 +18,11 @@ interface Props {
 
 export const ChooseProductModal: React.FC<Props> = ({ className, product }) => {
   const router = useRouter();
-  const firstItem = product.items[0];
-  const isPizzaForm = Boolean(firstItem.pizzaType);
-  const [addCartItem, loading] = useCartStore((state) => [state.addCartItem, state.loading]);
-
-  // const onAddProduct = async () => {
-  //   try {
-  //     await addCartItem({ productItemId: firstItem.id });
-  //     toast.success("Added product to cart");
-  //     router.back();
-  //   } catch (error) {
-  //     toast.error("Something went wrong");
-  //     console.error(error);
-  //   }
-  // };
-
-  // const onAddPizza = async (productItemId: number, ingredients: number[]) => {
-  //   try {
-  //     await addCartItem({ productItemId, ingredients });
-  //     toast.success("Added pizza to cart");
-  //     router.back();
-  //   } catch (error) {
-  //     toast.error("Something went wrong");
-  //     console.error(error);
-  //   }
-  // };
-
-  const onSubmit = async (productItemId?: number, ingredients?: number[]) => {
-    try {
-      if (isPizzaForm) {
-        await addCartItem({ productItemId, ingredients });
-      } else {
-        await addCartItem({ productItemId: firstItem.id });
-      }
-
-      toast.success(`${product.name} добавлен в корзину`);
-      router.back();
-    } catch (error) {
-      toast.error("Упс, что-то пошло не так");
-      console.error(error);
-    }
-  };
 
   return (
     <Dialog open={Boolean(product)} onOpenChange={() => router.back()}>
       <DialogContent className={cn("p-0 w-[1060px] max-w-[1060px] min-h-[600px] bg-white overflow-hidden", className)}>
-        {isPizzaForm ? (
-          <ChoosePizzaForm
-            imageUrl={product.imageUrl}
-            name={product.name}
-            ingredients={product.ingredients}
-            items={product.items}
-            onSubmit={onSubmit}
-            loading={loading}
-          />
-        ) : (
-          <ChooseProductForm
-            imageUrl={product.imageUrl}
-            name={product.name}
-            onSubmit={() => onSubmit()}
-            price={firstItem.price}
-            loading={loading}
-          />
-        )}
+        <ProductForm product={product} onSubmit={() => router.back()} />
       </DialogContent>
     </Dialog>
   );
