+++ shared/components/shared/product-form.tsx
@@ -1,0 +1,60 @@
+"use client";
+
+import React from "react";
+import toast from "react-hot-toast";
+import { useCartStore } from "@/shared/store/cart";
+import { ProductWithRelations } from "@/@types/prisma";
+import { ChoosePizzaForm } from "./choose-pizza-form";
+import { ChooseProductForm } from "./choose-product-form";
+
+interface Props {
+  product: ProductWithRelations;
+  onSubmit?: VoidFunction;
+}
+
+export const ProductForm: React.FC<Props> = ({ product, onSubmit: _onSubmit }) => {
+  const [addCartItem, loading] = useCartStore((state) => [state.addCartItem, state.loading]);
+
+  const firstItem = product.items[0];
+  const isPizzaForm = Boolean(firstItem.pizzaType);
+
+  const onSubmit = async (productItemId?: number, ingredients?: number[]) => {
+    try {
+      if (isPizzaForm) {
+        await addCartItem({ productItemId, ingredients });
+      } else {
+        await addCartItem({ productItemId: firstItem.id });
+      }
+
+      toast.success(`${product.name} добавлен в корзину`);
+
+      _onSubmit?.();
+    } catch (error) {
+      toast.error("Упс, что-то пошло не так");
+      console.error(error);
+    }
+  };
+
+  if (isPizzaForm) {
+    return (
+      <ChoosePizzaForm
+        imageUrl={product.imageUrl}
+        name={product.name}
+        ingredients={product.ingredients}
+        items={product.items}
+        onSubmit={onSubmit}
+        loading={loading}
+      />
+    );
+  }
+
+  return (
+    <ChooseProductForm
+      imageUrl={product.imageUrl}
+      name={product.name}
+      onSubmit={() => onSubmit()}
+      price={firstItem.price}
+      loading={loading}
+    />
+  );
+};
