+++ shared/lib/get-cart-details.ts
@@ -7,6 +7,7 @@ export type CartStateItem = {
   name: string;
   imageUrl: string;
   price: number;
+  disabled?: boolean;
   pizzaSize?: number | null;
   pizzaType?: number | null;
   ingredients: Array<{ name: string; price: number }>;
@@ -24,6 +25,7 @@ export const getCartDetails = (data: CartDTO) => {
       name: item.productItem.product.name,
       imageUrl: item.productItem.product.imageUrl,
       price: calcCartItemTotalPrice(item),
+      disabled: false,
       pizzaSize: item.productItem.size,
       pizzaType: item.productItem.pizzaType,
       ingredients: item.ingredients.map((ingredient) => ({
@@ -31,7 +33,7 @@ export const getCartDetails = (data: CartDTO) => {
         price: ingredient.price,
       })),
     };
-  });
+  }) as CartStateItem[];
 
   return {
     items,
