+++ shared/store/cart.ts
@@ -67,14 +67,18 @@ export const useCartStore = create<CartState>((set, get) => ({
   // DELETE /api/cart/:id
   removeCartItem: async (id: number) => {
     try {
-      set({ loading: true, error: false });
+      set((state) => ({
+        loading: true,
+        error: false,
+        items: state.items.map((item) => (item.id === id ? { ...item, disabled: true } : item)),
+      }));
       const data = await Api.cart.removeCartItem(id);
       set(getCartDetails(data));
     } catch (error) {
       set({ error: true });
       console.error(error);
     } finally {
-      set({ loading: false });
+      set((state) => ({ loading: false, items: state.items.map((item) => ({ ...item, disabled: false })) }));
     }
   },
